[#224] refactor - RouteFindingCameraViewController PHPickerViewControllerDelegate Extension 분리
